pyproject.toml
[project]
name = "personal-cli-assistant"
version = "0.1.0"
description = "A command-line AI assistant with function calling capabilities"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.3.0",
    "duckduckgo-search>=6.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
assistant = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
